@@ -1,12 +1,14 @@
 module github.com/lanrat/dns
 
-go 1.19
+go 1.23.0
+
+toolchain go1.24.2
 
 require (
-	golang.org/x/net v0.22.0
-	golang.org/x/sync v0.6.0
-	golang.org/x/sys v0.18.0
-	golang.org/x/tools v0.19.0
+	golang.org/x/net v0.42.0
+	golang.org/x/sync v0.16.0
+	golang.org/x/sys v0.34.0
+	golang.org/x/tools v0.35.0
 )
 
-require golang.org/x/mod v0.16.0 // indirect
+require golang.org/x/mod v0.26.0 // indirect
