@@ -14,7 +14,7 @@ import (
 // SVCBKey is the type of the keys used in the SVCB RR.
 type SVCBKey uint16
 
-// Keys defined in draft-ietf-dnsop-svcb-https-08 Section 14.3.2.
+// Keys defined in rfc9460
 const (
 	SVCB_MANDATORY SVCBKey = iota
 	SVCB_ALPN
@@ -23,7 +23,8 @@ const (
 	SVCB_IPV4HINT
 	SVCB_ECHCONFIG
 	SVCB_IPV6HINT
-	SVCB_DOHPATH // draft-ietf-add-svcb-dns-02 Section 9
+	SVCB_DOHPATH // rfc9461 Section 5
+	SVCB_OHTTP   // rfc9540 Section 8
 
 	svcb_RESERVED SVCBKey = 65535
 )
@@ -37,6 +38,7 @@ var svcbKeyToStringMap = map[SVCBKey]string{
 	SVCB_ECHCONFIG:       "ech",
 	SVCB_IPV6HINT:        "ipv6hint",
 	SVCB_DOHPATH:         "dohpath",
+	SVCB_OHTTP:           "ohttp",
 }
 
 var svcbStringToKeyMap = reverseSVCBKeyMap(svcbKeyToStringMap)
@@ -201,6 +203,8 @@ func makeSVCBKeyValue(key SVCBKey) SVCBKeyValue {
 		return new(SVCBIPv6Hint)
 	case SVCB_DOHPATH:
 		return new(SVCBDoHPath)
+	case SVCB_OHTTP:
+		return new(SVCBOhttp)
 	case svcb_RESERVED:
 		return nil
 	default:
@@ -210,11 +214,7 @@ func makeSVCBKeyValue(key SVCBKey) SVCBKeyValue {
 	}
 }
 
-// SVCB RR. See RFC xxxx (https://tools.ietf.org/html/draft-ietf-dnsop-svcb-https-08).
-//
-// NOTE: The HTTPS/SVCB RFCs are in the draft stage.
-// The API, including constants and types related to SVCBKeyValues, may
-// change in future versions in accordance with the latest drafts.
+// SVCB RR. See RFC 9460.
 type SVCB struct {
 	Hdr      RR_Header
 	Priority uint16         // If zero, Value must be empty or discarded by the user of this library
@@ -222,12 +222,8 @@ type SVCB struct {
 	Value    []SVCBKeyValue `dns:"pairs"`
 }
 
-// HTTPS RR. Everything valid for SVCB applies to HTTPS as well.
+// HTTPS RR. See RFC 9460. Everything valid for SVCB applies to HTTPS as well.
 // Except that the HTTPS record is intended for use with the HTTP and HTTPS protocols.
-//
-// NOTE: The HTTPS/SVCB RFCs are in the draft stage.
-// The API, including constants and types related to SVCBKeyValues, may
-// change in future versions in accordance with the latest drafts.
 type HTTPS struct {
 	SVCB
 }
@@ -302,7 +298,7 @@ func (s *SVCBMandatory) pack() ([]byte, error) {
 
 func (s *SVCBMandatory) unpack(b []byte) error {
 	if len(b)%2 != 0 {
-		return errors.New("dns: svcbmandatory: value length is not a multiple of 2")
+		return errors.New("bad svcbmandatory: value length is not a multiple of 2")
 	}
 	codes := make([]SVCBKey, 0, len(b)/2)
 	for i := 0; i < len(b); i += 2 {
@@ -399,10 +395,10 @@ func (s *SVCBAlpn) pack() ([]byte, error) {
 	b := make([]byte, 0, 10*len(s.Alpn))
 	for _, e := range s.Alpn {
 		if e == "" {
-			return nil, errors.New("dns: svcbalpn: empty alpn-id")
+			return nil, errors.New("bad svcbalpn: empty alpn-id")
 		}
 		if len(e) > 255 {
-			return nil, errors.New("dns: svcbalpn: alpn-id too long")
+			return nil, errors.New("bad svcbalpn: alpn-id too long")
 		}
 		b = append(b, byte(len(e)))
 		b = append(b, e...)
@@ -417,7 +413,7 @@ func (s *SVCBAlpn) unpack(b []byte) error {
 		length := int(b[i])
 		i++
 		if i+length > len(b) {
-			return errors.New("dns: svcbalpn: alpn array overflowing")
+			return errors.New("bad svcbalpn: alpn array overflowing")
 		}
 		alpn = append(alpn, string(b[i:i+length]))
 		i += length
@@ -437,13 +433,13 @@ func (s *SVCBAlpn) parse(b string) error {
 	for p := 0; p < len(b); {
 		c, q := nextByte(b, p)
 		if q == 0 {
-			return errors.New("dns: svcbalpn: unterminated escape")
+			return errors.New("bad svcbalpn: unterminated escape")
 		}
 		p += q
 		// If we find a comma, we have finished reading an alpn.
 		if c == ',' {
 			if len(a) == 0 {
-				return errors.New("dns: svcbalpn: empty protocol identifier")
+				return errors.New("bad svcbalpn: empty protocol identifier")
 			}
 			alpn = append(alpn, string(a))
 			a = []byte{}
@@ -453,10 +449,10 @@ func (s *SVCBAlpn) parse(b string) error {
 		if c == '\\' {
 			dc, dq := nextByte(b, p)
 			if dq == 0 {
-				return errors.New("dns: svcbalpn: unterminated escape decoding comma-separated list")
+				return errors.New("bad svcbalpn: unterminated escape decoding comma-separated list")
 			}
 			if dc != '\\' && dc != ',' {
-				return errors.New("dns: svcbalpn: bad escaped character decoding comma-separated list")
+				return errors.New("bad svcbalpn: bad escaped character decoding comma-separated list")
 			}
 			p += dq
 			c = dc
@@ -465,7 +461,7 @@ func (s *SVCBAlpn) parse(b string) error {
 	}
 	// Add the final alpn.
 	if len(a) == 0 {
-		return errors.New("dns: svcbalpn: last protocol identifier empty")
+		return errors.New("bad svcbalpn: last protocol identifier empty")
 	}
 	s.Alpn = append(alpn, string(a))
 	return nil
@@ -503,14 +499,14 @@ func (*SVCBNoDefaultAlpn) len() int              { return 0 }
 
 func (*SVCBNoDefaultAlpn) unpack(b []byte) error {
 	if len(b) != 0 {
-		return errors.New("dns: svcbnodefaultalpn: no-default-alpn must have no value")
+		return errors.New("bad svcbnodefaultalpn: no-default-alpn must have no value")
 	}
 	return nil
 }
 
 func (*SVCBNoDefaultAlpn) parse(b string) error {
 	if b != "" {
-		return errors.New("dns: svcbnodefaultalpn: no-default-alpn must have no value")
+		return errors.New("bad svcbnodefaultalpn: no-default-alpn must have no value")
 	}
 	return nil
 }
@@ -533,7 +529,7 @@ func (s *SVCBPort) copy() SVCBKeyValue { return &SVCBPort{s.Port} }
 
 func (s *SVCBPort) unpack(b []byte) error {
 	if len(b) != 2 {
-		return errors.New("dns: svcbport: port length is not exactly 2 octets")
+		return errors.New("bad svcbport: port length is not exactly 2 octets")
 	}
 	s.Port = binary.BigEndian.Uint16(b)
 	return nil
@@ -548,7 +544,7 @@ func (s *SVCBPort) pack() ([]byte, error) {
 func (s *SVCBPort) parse(b string) error {
 	port, err := strconv.ParseUint(b, 10, 16)
 	if err != nil {
-		return errors.New("dns: svcbport: port out of range")
+		return errors.New("bad svcbport: port out of range")
 	}
 	s.Port = uint16(port)
 	return nil
@@ -581,7 +577,7 @@ func (s *SVCBIPv4Hint) pack() ([]byte, error) {
 	for _, e := range s.Hint {
 		x := e.To4()
 		if x == nil {
-			return nil, errors.New("dns: svcbipv4hint: expected ipv4, hint is ipv6")
+			return nil, errors.New("bad svcbipv4hint: expected ipv4, hint is ipv6")
 		}
 		b = append(b, x...)
 	}
@@ -590,7 +586,7 @@ func (s *SVCBIPv4Hint) pack() ([]byte, error) {
 
 func (s *SVCBIPv4Hint) unpack(b []byte) error {
 	if len(b) == 0 || len(b)%4 != 0 {
-		return errors.New("dns: svcbipv4hint: ipv4 address byte array length is not a multiple of 4")
+		return errors.New("bad svcbipv4hint: ipv4 address byte array length is not a multiple of 4")
 	}
 	b = cloneSlice(b)
 	x := make([]net.IP, 0, len(b)/4)
@@ -615,10 +611,10 @@ func (s *SVCBIPv4Hint) String() string {
 
 func (s *SVCBIPv4Hint) parse(b string) error {
 	if b == "" {
-		return errors.New("dns: svcbipv4hint: empty hint")
+		return errors.New("bad svcbipv4hint: empty hint")
 	}
 	if strings.Contains(b, ":") {
-		return errors.New("dns: svcbipv4hint: expected ipv4, got ipv6")
+		return errors.New("bad svcbipv4hint: expected ipv4, got ipv6")
 	}
 
 	hint := make([]net.IP, 0, strings.Count(b, ",")+1)
@@ -627,7 +623,7 @@ func (s *SVCBIPv4Hint) parse(b string) error {
 		e, b, _ = strings.Cut(b, ",")
 		ip := net.ParseIP(e).To4()
 		if ip == nil {
-			return errors.New("dns: svcbipv4hint: bad ip")
+			return errors.New("bad svcbipv4hint: bad ip")
 		}
 		hint = append(hint, ip)
 	}
@@ -675,7 +671,7 @@ func (s *SVCBECHConfig) unpack(b []byte) error {
 func (s *SVCBECHConfig) parse(b string) error {
 	x, err := fromBase64([]byte(b))
 	if err != nil {
-		return errors.New("dns: svcbech: bad base64 ech")
+		return errors.New("bad svcbech: bad base64 ech")
 	}
 	s.ECH = x
 	return nil
@@ -703,7 +699,7 @@ func (s *SVCBIPv6Hint) pack() ([]byte, error) {
 	b := make([]byte, 0, 16*len(s.Hint))
 	for _, e := range s.Hint {
 		if len(e) != net.IPv6len || e.To4() != nil {
-			return nil, errors.New("dns: svcbipv6hint: expected ipv6, hint is ipv4")
+			return nil, errors.New("bad svcbipv6hint: expected ipv6, hint is ipv4")
 		}
 		b = append(b, e...)
 	}
@@ -712,14 +708,14 @@ func (s *SVCBIPv6Hint) pack() ([]byte, error) {
 
 func (s *SVCBIPv6Hint) unpack(b []byte) error {
 	if len(b) == 0 || len(b)%16 != 0 {
-		return errors.New("dns: svcbipv6hint: ipv6 address byte array length not a multiple of 16")
+		return errors.New("bas svcbipv6hint: ipv6 address byte array length not a multiple of 16")
 	}
 	b = cloneSlice(b)
 	x := make([]net.IP, 0, len(b)/16)
 	for i := 0; i < len(b); i += 16 {
 		ip := net.IP(b[i : i+16])
 		if ip.To4() != nil {
-			return errors.New("dns: svcbipv6hint: expected ipv6, got ipv4")
+			return errors.New("bad svcbipv6hint: expected ipv6, got ipv4")
 		}
 		x = append(x, ip)
 	}
@@ -740,7 +736,7 @@ func (s *SVCBIPv6Hint) String() string {
 
 func (s *SVCBIPv6Hint) parse(b string) error {
 	if b == "" {
-		return errors.New("dns: svcbipv6hint: empty hint")
+		return errors.New("bad svcbipv6hint: empty hint")
 	}
 
 	hint := make([]net.IP, 0, strings.Count(b, ",")+1)
@@ -749,10 +745,10 @@ func (s *SVCBIPv6Hint) parse(b string) error {
 		e, b, _ = strings.Cut(b, ",")
 		ip := net.ParseIP(e)
 		if ip == nil {
-			return errors.New("dns: svcbipv6hint: bad ip")
+			return errors.New("bad svcbipv6hint: bad ip")
 		}
 		if ip.To4() != nil {
-			return errors.New("dns: svcbipv6hint: expected ipv6, got ipv4-mapped-ipv6")
+			return errors.New("bad svcbipv6hint: expected ipv6, got ipv4-mapped-ipv6")
 		}
 		hint = append(hint, ip)
 	}
@@ -771,8 +767,8 @@ func (s *SVCBIPv6Hint) copy() SVCBKeyValue {
 // SVCBDoHPath pair is used to indicate the URI template that the
 // clients may use to construct a DNS over HTTPS URI.
 //
-// See RFC xxxx (https://datatracker.ietf.org/doc/html/draft-ietf-add-svcb-dns-02)
-// and RFC yyyy (https://datatracker.ietf.org/doc/html/draft-ietf-add-ddr-06).
+// See RFC 9461 (https://datatracker.ietf.org/doc/html/rfc9461)
+// and RFC 9462 (https://datatracker.ietf.org/doc/html/rfc9462).
 //
 // A basic example of using the dohpath option together with the alpn
 // option to indicate support for DNS over HTTPS on a certain path:
@@ -804,7 +800,7 @@ func (s *SVCBDoHPath) unpack(b []byte) error {
 func (s *SVCBDoHPath) parse(b string) error {
 	template, err := svcbParseParam(b)
 	if err != nil {
-		return fmt.Errorf("dns: svcbdohpath: %w", err)
+		return fmt.Errorf("bad svcbdohpath: %w", err)
 	}
 	s.Template = string(template)
 	return nil
@@ -814,6 +810,44 @@ func (s *SVCBDoHPath) copy() SVCBKeyValue {
 	return &SVCBDoHPath{
 		Template: s.Template,
 	}
+}
+
+// The "ohttp" SvcParamKey is used to indicate that a service described in a SVCB RR
+// can be accessed as a target using an associated gateway.
+// Both the presentation and wire-format values for the "ohttp" parameter MUST be empty.
+//
+// See RFC 9460 (https://datatracker.ietf.org/doc/html/rfc9460/)
+// and RFC 9230 (https://datatracker.ietf.org/doc/html/rfc9230/)
+//
+// A basic example of using the dohpath option together with the alpn
+// option to indicate support for DNS over HTTPS on a certain path:
+//
+//	s := new(dns.SVCB)
+//	s.Hdr = dns.RR_Header{Name: ".", Rrtype: dns.TypeSVCB, Class: dns.ClassINET}
+//	e := new(dns.SVCBAlpn)
+//	e.Alpn = []string{"h2", "h3"}
+//	p := new(dns.SVCBOhttp)
+//	s.Value = append(s.Value, e, p)
+type SVCBOhttp struct{}
+
+func (*SVCBOhttp) Key() SVCBKey          { return SVCB_OHTTP }
+func (*SVCBOhttp) copy() SVCBKeyValue    { return &SVCBOhttp{} }
+func (*SVCBOhttp) pack() ([]byte, error) { return []byte{}, nil }
+func (*SVCBOhttp) String() string        { return "" }
+func (*SVCBOhttp) len() int              { return 0 }
+
+func (*SVCBOhttp) unpack(b []byte) error {
+	if len(b) != 0 {
+		return errors.New("bad svcbotthp: svcbotthp must have no value")
+	}
+	return nil
+}
+
+func (*SVCBOhttp) parse(b string) error {
+	if b != "" {
+		return errors.New("bad svcbotthp: svcbotthp must have no value")
+	}
+	return nil
 }
 
 // SVCBLocal pair is intended for experimental/private use. The key is recommended
@@ -844,7 +878,7 @@ func (s *SVCBLocal) unpack(b []byte) error {
 func (s *SVCBLocal) parse(b string) error {
 	data, err := svcbParseParam(b)
 	if err != nil {
-		return fmt.Errorf("dns: svcblocal: svcb private/experimental key %w", err)
+		return fmt.Errorf("bad svcblocal: svcb private/experimental key %w", err)
 	}
 	s.Data = data
 	return nil
